Changed data type for handler context data
Previously, the data type was a HashMap; however, we want to do things
in the future that require backend data to be loaded in a predictable
order, and since HashMap is unordered, it won't do.

A plain association list [(Text, Value)] would do, but it would be
incompatible with the existing YAML file format for configurations and
projects, and it would serialize to a rather clumsy format of
`[[key, value], ...]`.

Instead, we're introducing a newtyped association list (`AList`), which
serializes to a list of singleton objects `[{key: value}, {key:
value}]`, which makes for nice YAML output, and deserializes from
objects, lists of objects, lists of lists of objects, etc., to arbitrary
nesting levels, maintaining ordering as much as possible.

In practice, this means that we can write, interchangeably:

    - foo: bar
    - baz: quux

and

    foo: bar
    baz: quux

and

    - foo: bar
      baz: quux

...and they will all parse as the same data structure, *except* that
only the first one guarantees ordering.