Expose loaded backend data to backend specs defined later.
Backend data items are now added to the backend spec resolver context
for all backend specs defined later. This allows us to define backends
that depend on the output of other backends, e.g., we can have a
settings.yml that provides database connection information that we can
dereference in other data sources.