Regression: multi-item matches stringify as a/b/c
Multi-item path matches are now captures as lists of items rather than
pre-concatenated paths; however, if we stringify them in Ginger, we want
them to join on "/", not "" or " ". To fix this, we override the Ginger
asText method, which allows captured paths to behave like strings but
also like lists. Caveat is that once we treat them as lists and do
processing on them, the special asText behavior goes away, so
{{path[:2]}} on a captured path of "foo/bar/baz" would produce
"foobar", not "foo/bar", because the [:2] operator returns a vanilla
list.