Fix excessively long filenames in filesystem cache
Previously, we would construct filenames from a reversible encoding of
the cache key, but cache keys can grow awkwardly long, and with the
added encoding overhead, they can blow the filename length limit of the
underlying filesystem.

But it turns out we don't actually need to reverse the encoding, ever,
so we can simply hash the cache key with sha1 instead, which gives us a
constant filename length.