Fix pandoc code block conversion
Previously, pandoc Code blocks would incorrectly assume that their
contents is a list of inlines; however, it is actually a String, which,
due to String being an alias for [Char], means that converting it as a
list of things still "works" - but you get a list of individual
characters instead of a singleton list containing the whole string.

Ugh.